SASS-10232 - Fixed test and removed classes no longer needed
